src/agents/gemini_brain.py
"""
Gemini 2.0 Strategic Brain
Advanced AI decision making for OGame strategy
"""

import os
import google.generativeai as genai
from typing import Dict, List, Any, Optional
import logging


class GeminiBrain:
    """Strategic AI brain powered by Gemini 2.0"""
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.setup_gemini()
        
    def setup_gemini(self):
        """Initialize Gemini AI"""
        api_key = os.getenv('GEMINI_API_KEY')
        if not api_key:
            raise ValueError("GEMINI_API_KEY environment variable required")
            
        genai.configure(api_key=api_key)
        self.model = genai.GenerativeModel('gemini-2.0-flash-exp')
        self.logger.info("🧠 Gemini brain initialized")
        
    async def analyze_game_state(self, game_data: Dict[str, Any]) -> Dict[str, Any]:
        """Analyze current game state and provide strategic insights"""
        
        prompt = f"""
        You are an expert OGame strategist. Analyze this game state and provide strategic recommendations:
        
        Game Data: {game_data}
        
        Consider:
        1. Resource production efficiency
        2. Fleet composition optimization  
        3. Research priorities
        4. Expansion opportunities
        5. Defense requirements
        6. Risk assessment
        
        Respond with structured JSON containing:
        - immediate_actions: List of urgent actions needed
        - strategic_priorities: Long-term goals in order of importance
        - resource_allocation: Recommended resource distribution
        - risk_level: Current threat assessment (1-10)
        - recommendations: Specific actionable advice
        """
        
        try:
            response = await self.model.generate_content_async(prompt)
            # TODO: Parse and validate response
            return {"analysis": response.text}
        except Exception as e:
            self.logger.error(f"Error in game state analysis: {e}")
            return {"error": str(e)}
            
    async def make_decision(self, situation: str, options: List[str]) -> str:
        """Make strategic decisions based on current situation"""
        
        prompt = f"""
        OGame Strategic Decision Required:
        
        Situation: {situation}
        Available Options: {options}
        
        As an expert OGame player, choose the best option and explain your reasoning.
        Consider efficiency, risk, and strategic value.
        
        Respond with just the chosen option and brief reasoning.
        """
        
        try:
            response = await self.model.generate_content_async(prompt)
            return response.text.strip()
        except Exception as e:
            self.logger.error(f"Error in decision making: {e}")
            return options[0] if options else "No action"
            
    async def analyze_screenshot(self, image_data: bytes) -> Dict[str, Any]:
        """Analyze game screenshot using Gemini vision capabilities"""
        
        # TODO: Implement vision analysis
        self.logger.info("🔍 Analyzing screenshot with Gemini vision")
        return {"vision_analysis": "Not implemented yet"}