src/main.py
"""
OGame AGI - Main Entry Point
Autonomous Game Intelligence for OGame
"""

import asyncio
import logging
from pathlib import Path
from dotenv import load_dotenv

from agents.gemini_brain import GeminiBrain
from automation.playwright_interface import PlaywrightInterface
from strategy.strategic_planner import StrategicPlanner
from vision.game_parser import GameParser


class OGameAGI:
    """Main OGame AGI Agent Controller"""
    
    def __init__(self):
        load_dotenv()
        self.setup_logging()
        
        # Initialize components
        self.brain = GeminiBrain()
        self.browser = PlaywrightInterface()
        self.parser = GameParser()
        self.planner = StrategicPlanner(self.brain)
        
        self.running = False
        
    def setup_logging(self):
        """Configure logging system"""
        logging.basicConfig(
            level=logging.INFO,
            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
            handlers=[
                logging.FileHandler('/app/logs/ogame-agi.log'),
                logging.StreamHandler()
            ]
        )
        self.logger = logging.getLogger(__name__)
        
    async def start(self):
        """Start the OGame AGI agent"""
        self.logger.info("🚀 Starting OGame AGI Agent...")
        
        try:
            # Initialize browser
            await self.browser.initialize()
            self.logger.info("✅ Browser initialized")
            
            # TODO: Add game login and main loop
            self.running = True
            await self.main_loop()
            
        except Exception as e:
            self.logger.error(f"❌ Error starting agent: {e}")
        finally:
            await self.cleanup()
            
    async def main_loop(self):
        """Main agent execution loop"""
        while self.running:
            try:
                # TODO: Implement main game logic
                self.logger.info("🔄 Agent cycle...")
                await asyncio.sleep(10)  # Placeholder
                
            except KeyboardInterrupt:
                self.logger.info("🛑 Shutdown requested")
                self.running = False
            except Exception as e:
                self.logger.error(f"❌ Error in main loop: {e}")
                await asyncio.sleep(5)
                
    async def cleanup(self):
        """Cleanup resources"""
        self.logger.info("🧹 Cleaning up resources...")
        await self.browser.close()
        self.logger.info("✅ Cleanup complete")


async def main():
    """Application entry point"""
    agent = OGameAGI()
    await agent.start()


if __name__ == "__main__":
    asyncio.run(main())