src/vision/game_parser.py
"""
Game Parser - Computer Vision for OGame Interface
Extract game data from UI elements and screenshots
"""

import cv2
import numpy as np
from PIL import Image
import logging
from typing import Dict, Any, List, Optional, Tuple
import io


class GameParser:
    """Computer vision parser for OGame interface"""
    
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        
    def parse_screenshot(self, screenshot_data: bytes) -> Dict[str, Any]:
        """Parse game screenshot and extract relevant data"""
        try:
            # Convert bytes to OpenCV image
            image = self._bytes_to_cv_image(screenshot_data)
            
            parsed_data = {
                'resources': self._extract_resources(image),
                'fleet_info': self._extract_fleet_info(image),
                'planet_info': self._extract_planet_info(image),
                'research_info': self._extract_research_info(image),
                'messages': self._extract_messages(image)
            }
            
            self.logger.info("🔍 Screenshot parsed successfully")
            return parsed_data
            
        except Exception as e:
            self.logger.error(f"Screenshot parsing failed: {e}")
            return {}
            
    def _bytes_to_cv_image(self, image_bytes: bytes) -> np.ndarray:
        """Convert image bytes to OpenCV format"""
        image = Image.open(io.BytesIO(image_bytes))
        return cv2.cvtColor(np.array(image), cv2.COLOR_RGB2BGR)
        
    def _extract_resources(self, image: np.ndarray) -> Dict[str, int]:
        """Extract resource counts from the interface"""
        # TODO: Implement OCR for resource numbers
        # This would involve:
        # 1. Locating resource display area
        # 2. OCR to read numbers
        # 3. Parsing metal, crystal, deuterium values
        
        return {
            'metal': 0,
            'crystal': 0, 
            'deuterium': 0,
            'energy': 0
        }
        
    def _extract_fleet_info(self, image: np.ndarray) -> Dict[str, Any]:
        """Extract fleet information"""
        # TODO: Parse fleet composition and status
        return {
            'total_ships': 0,
            'fleet_slots_used': 0,
            'fleet_slots_total': 0,
            'expeditions': []
        }
        
    def _extract_planet_info(self, image: np.ndarray) -> Dict[str, Any]:
        """Extract planet information"""
        # TODO: Parse planet details
        return {
            'name': '',
            'coordinates': '',
            'fields_used': 0,
            'fields_total': 0,
            'temperature': {'min': 0, 'max': 0}
        }
        
    def _extract_research_info(self, image: np.ndarray) -> Dict[str, Any]:
        """Extract research laboratory information"""
        # TODO: Parse research status
        return {
            'current_research': None,
            'research_queue': [],
            'completion_time': None
        }
        
    def _extract_messages(self, image: np.ndarray) -> List[Dict[str, Any]]:
        """Extract messages and notifications"""
        # TODO: Parse message indicators
        return []
        
    def find_ui_elements(self, image: np.ndarray, element_type: str) -> List[Tuple[int, int, int, int]]:
        """Find UI elements using template matching or feature detection"""
        # TODO: Implement template matching for buttons, menus, etc.
        return []
        
    def extract_text_from_region(self, image: np.ndarray, bbox: Tuple[int, int, int, int]) -> str:
        """Extract text from specific region using OCR"""
        # TODO: Implement OCR using pytesseract or similar
        x, y, w, h = bbox
        region = image[y:y+h, x:x+w]
        # OCR processing would go here
        return ""
        
    def detect_game_state(self, image: np.ndarray) -> str:
        """Detect current game screen/state"""
        # TODO: Classify current screen (overview, fleetdispatch, research, etc.)
        return "overview"