requirements.txt
# AI and Web Automation
playwright==1.47.0
google-generativeai==0.8.3

# Computer Vision
opencv-python==4.10.0.84
pillow==10.4.0

# Web Scraping & HTTP
requests==2.32.3
beautifulsoup4==4.12.3

# Data Processing
numpy==1.26.4
pandas==2.2.3

# Async Programming
asyncio==3.4.3
aiohttp==3.10.5

# Configuration
python-dotenv==1.0.1
pydantic==2.9.2

# Logging and Monitoring
loguru==0.7.2
rich==13.8.1

# Testing (development)
pytest==8.3.3
pytest-asyncio==0.24.0